+++ crates/rattler-bin/src/commands/mod.rs
@@ -15,6 +15,7 @@ pub mod progress;
 pub mod run;
 pub mod search;
 pub mod shell_hook;
+pub mod solve;
 pub mod virtual_packages;
 
 #[cfg(feature = "sigstore")]